Use new device class in meteo_france
